split code in modules and remove duplicate code
